+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Bugfixes
 
+* Fixed [#735](https://github.com/appserver-io/appserver/issues/735) - Endless Loop for URLs without servlet name
 * Fixed [#719](https://github.com/appserver-io/appserver/issues/719) - Around advice chain does break at certain size
 * Fixed [#721](https://github.com/appserver-io/appserver/issues/721) - Different order of Advices in pointcut.xml depending on type
 
@@ -36,6 +37,8 @@
 
 ## Bugfixes
 
+* Fixed [#725](https://github.com/appserver-io/appserver/issues/725) - no Datasources in Singleton SessionBean
+* Fixed [#731](https://github.com/appserver-io/appserver/issues/731) - Custom include paths in SplClassLoader not used
 * Fixed [#719](https://github.com/appserver-io/appserver/issues/719) - Around advice chain does break at certain size
 * Fixed [#721](https://github.com/appserver-io/appserver/issues/721) - Different order of Advices in pointcut.xml depending on type